Use an empty string as default CustomName in settings instead of ProductName.
As long as a user does not specify a custom name for the device, the 
product name will be shown. The product name used to be the default
value in the local settings. This is not necessary and may get in the way
when we device to change the product name, because custom name
(in local settings) may still contain the old product name.

If CustomName in the localsettings is empty, the entry in the 
com.victronenergy.pvinverter tree will contain the current ProductName.