Do not use modbus TCP if inverter is not in int+sf mode
The Fronius modbus TCP protocol specifies 2 modes: float and int+sf.
If the float mode is active, some registers will be encoded as
floats (for example, current power). In int+sf mode, those values
are encoded as integers.

We support only int+sf mode. If the inverter is in float mode, it is
still possible to use the int+sf registers, but their values have
different meanings, so the data published on the D-Bus is incorrect.

To fix this, we check the mode and do not retrieve and publish
modbus values as long as the inverter is in float mode. If the mode
is changed to int+sf (using the web-interface on the inverter), this
will be detected and modbus communication will be started again.

The detect changes in the mode setting, we will retrieve the model
type register (216) from time to time.