Remove always notify feature.
Some D-BUS object (like the power limit) would always send a
PropertiesChanged signal in reaction to a SetValue. Even if the
value did not change.

This was intended as a fix for the storage problem in the VeQItems.
(calling VeQItem::setValue multiple times with the same value would
not cause a SetValue on the D-Bus). Normally this is a good thing,
unless the SetValue should be sent on regular intervals (keep-alive).

Because dbus-fronius breaks the Victron D-Bus standard (a little)
by always sending the PropertiesChanged, this is not the right place
to fix this. The new fix has been implemented in hub4control (which
sends the SetValue).