Made the scan progress percentage more accurate.
In previous versions the scan percentage would be calculated
from the number of IP addresses retrieved from the
LocalIpAddressGenerator class. However, some of those addresses
may still be under investigation (ie. we are still waiting for
a response from that address). This caused the scan progress
to reach 100% before all scans were completed.

Now, the percentage is corrected (downwards) for the addresses
still under investigation.