Revert "Qt5 compatibility"
Commit 7fd7c4e107d8aa33 "Use QHttp (Obsolete!) instead of QNetworkManager"
replaced the QNetworkManager by the deprecated QHttp class since as it was
causing problems: "Fixed a crash when scanning for PV inverters on a large
network (/8 or /16 netmask)."

Given that the docs says "One QNetworkAccessManager should be enough for
the whole Qt application." it is not suprising that it result in odd
behaviour when creating it per host. E.g. it will spawn 32 threads to
wait for requests instead of using a single select. In order to make sure
we don't get the same problems again as we had years ago, lets postpone
this for now.

This reverts commit 3b5ee3cdbb96fbd433a222bc0afcb8b7f02156db.