solar_api_updater: don't set parent of member
Setting a parent will normally deallocate the object when the parent is
destructed, so when setting a parent it looks as if non allocated will
be freed. This doesn't happen, since:
 - the parent destructor runs first
 - the parents member destructor will run (setting parent to 0)
 - the qObject destructor will run and delete the children

So while it doesn't crash, it is confusing to read, so lets simply not
set the parent in the first place.