Add support for a single phase inverter across a split-phase installation.
This is typical in North America. A Single-phase 208/240V PV-inverter might
be connected across a split-phase (240V, pi radians) or a two-out-of-three
phases (208V, 2pi/3 radians) connection.

If the PV-inverter is configured as Multi-phase while its sunspec model
only indicates a single phase, we assume that the power and energy is
split equally over the two phases. Per-phase voltage is not known and not
reported, and though the current is known it is not reported to avoid
confusion.