Add AutoScan setting to disable automatic scan
On startup dbus-fronius will always attempt to scan for PV-inverters found a
previous time. If they are not found, or if no PV-inverters have ever been
found, it will always launch a full scan of the whole subnet. This makes
it possible to disable that behaviour.

This was requested by a customer where the scanning has detrimental effects
to other equipment on the network.