Persist Inverter's PhaseCount to localsettings.
This allows the GUI to do the right thing: If the PV-inverter is multi-phase
dbus-fronius will already ignore the Phase setting, so the GUI should also
hide the phase selection option (it does this already, but based on different
logic).

If the PV-inverter is single-phase it can however still operate across phases
in a split-phase system, so the GUI should show the single-phase options in
addition to a multi-phase option.

The GUI can now adjust according to the actual phase count rather than the
multi-phase configuration setting, since multi-phase is possible with a
single-phase PV-inverter.

https://github.com/victronenergy/venus/issues/458