Avoid unnecessary rescanning when inverter found
The signals for acting on changes to inverter-specific settings were
attached too early, causing changed events to fire when we connect to the
setting items.  Since we do this when an inverter is detected, and the
change-handler would immediately start a new scan, this caused an immediate
restart to scanning when an inverter is found.

Now we delay attaching these settings until after the settings are
initialised, so that it only reacts to future changes.