Don't mix manually configured and autodetected ips
dbus-fronius remembers the addresses of detected PV inverters and
stores it in localsettings. It however used to also duplicate manually
configured addresses into this storage as if they were auto-detected.
Of course this means that any address, once manually configured, is
always remembered and always tried in future, but if users are allowed
to delete manually configured addresses on the GUI it makes sense that
we also stop scanning for them here.