Only do one gratuitous full scan after reboot
When not all the hosts are found during a priority scan, a full scan is done
in case the PV-inverter has a new IP address assigned. This causes an
inifinite loop looking for PV-inverters on their old addresses, so that
a full scan is launched every minute. Limit this to once after startup,
so that after a reboot a PV-inverter will be found in its new location,
but we are no longer permanently scanning for it.