Do periodic UDP scan (if permitted)
When doing a periodic scan of known hosts, also do a fast UDP scan
(if permitted). This could lead to finding the PV-inverter without having
to switch to a Full scan, eg when IP address changed. It also means new
PV-inverters are found automatically, as long as automatic scanning is on.