increase sunspec scan timeout on Fronius DataManager
A heavily loaded DataManager (observed with 6 PV-inverters) may become
very slow to scanning. Normally we don't want to make the timeout too
long since that will increase the time it takes to scan an entire subnet.
If we have however already determined the presence of a Fronius DataManger
we can increase the timeout for the sunspec scan to accommodate slow
DataManagers. This will not significantly slow down the overall scan, it
only adds max 10 seconds per discovered Fronius PV-inverter.