use uniqueId in localsettings if serial not available
Some Fronius PV-inverters cannot communicate the serial number
and for those we use the unique id (from solarapi) in its stead.
In order to keep this logic out of the GUI, persist the uniqueId
in localsettings if the serial number is not available. That way
the GUI need only look in one place.