fix incorrect handling of connection timeout
If there is a connection timeout to the PV inverter, this would
incorrectly disconnect the socket object from all its signal handlers.
This broke the entire data handling making it impossible to receive
data from a PV-inverter even if the connection was subsequently
established.

This affects particularly SolarEdge inverters that only allow
one concurrent connection and may be slow in accepting a new
connection after another was terminated.

https://github.com/victronenergy/venus/issues/760