Prevent switching back from sunspec to solarapi
Sunspec is the preferred and more capable protocol. It is however
possible that slow network conditions or a slow datamanager can cause
the detection process to time out on sunspec after succeeding on
solarapi, causing it to switch back and forth. To prevent this, refuse
to downgrade to solarapi, and instead let the normal timeout mechanism
deal with it. In other words, only allow the retrieval protocol to
change up, not down.

https://github.com/victronenergy/venus/issues/753