fix issue in sunspec scan optimisation 0d05a0
There are ways to lose a connection that does not
send the connectionLost signal. So instead, make
sure we remove the connection from the tracking
list when the Updater object itself is deleted,
which happens when the Inverter object is deleted
since the updater is a child to the inverter.