yet another fix on sunspec scan optimisation 0d05a0
If the IP address of the inverter changes, often because
it is connected both via LAN and Wi-Fi and it also finds
the other one, then after two changes it will lose the
connection permanently. To deal with this, keep track
of the updaters instead of the host+id pairs, so changes
to connection details are properly handled.