remove reading back of WMaxLimPct & WMaxLimPvt_Ena
Until now we would read back the values from model 123 after
writing a power limit. This caused, among others, the inability
to support the limiter on some SMA PV-inverters. But it also
affects newer Fimer PV-inverters (formerly ABB), which for
whatever reason keeps WMaxLimPvt_Ena at zero, making it
impossible for the existing algorithm to detect a successful
write.

So now we completely skip the read-back. We already know what
we wrote to it, and we simply assume the write operation
succeeded. This could potentially open the door in future to
limiting other brands as well.