add a power limit reset timer
In the past we could rely on the reversion timer in the PV-inverter
to indicate when the limit expired. Now we instead time out after
60 seconds (half the reversion counter) and reset it to full power
ourselves.

This commit also fixes a bug that went unnoticed for a long time.
The power percentage was stored as a float between 0 and 1, but
a in several places it was assumed to be between 0 and 100.