update solarapi devicetype detection list.
For SolarAPI, each device has its own id. The list is hardcoded
and mostly useless, as everyone should be using SunSpec instead.
But for cases where SolarAPI is still desired, this makes it work.

This also adds support for SolarAPI on the Tauro, but explicitly
breaks it for the Gen24, which identifies as the same device.

A Gen24 can still be used as a Sunspec device.

https://github.com/victronenergy/venus/issues/1022