Workaround for 24-bit uniqueId bug
In some cases -- exactly when this happens is unclear -- some or
all PV inverters come up with uniqueId 16777215, or 0xFFFFFF.

Historically we are stuck with uniqueId, because we don't know
for sure if all PV-inverters of all ages support fetching the
serial number through SolarAPI.

As a workaround for these broken Datamanagers, this adds a setting
which a support person (me) can toggle on such sites. It will
switch to using the serial number for identification.

https://github.com/victronenergy/venus/issues/1036