make sure dbus name is valid
The dbus name of a service contains the serial number. This means
the serial number must not contain characters that are not valid
in a dbus service name. For our purposes, this also means a dot
(.) because a dot creates another component name, and in the
case of some PV-inverters a digit may follow that dot, which is
also illegal.

Therefore, replace all characters that are not alphanumberic, an
underscore or a dash, with an underscore.

This fixes the problem for KACO inverters that has a dot in the
serial number, followed by a digit.

https://github.com/victronenergy/venus-private/issues/276