clean up storing namePlateModelOffset
This value is never used. Data from model 120, max power, is
immediately stored. There is no need to keep track of the offset
of model 120.

As a side effect, this will cause any PV-inverters that declare
a maxpower of zero to no longer be detected. This sometimes
happens with ABB inverters that incorrectly report no capacity.