Sunspec limiting: Move to separate class
This is so we can separate the limiting mechanism from the
power-and-voltage reading mechanism. In other words, this allows
reading power and voltage from a 100-series model while using
a 700-series model for limiting, and vice versa.

It also opens the door, in future, to use non-sunspec limiting
implementations for specific inverters.