cleanup of unused solarapi items
The errorCode, statusCode, and customName returned when scanning
inverters using solarApi was not used at all. No point in storing
them then.

Note about customName: This is distinct from the customName used
on the GX side. The customName can not be set on anything other
than the Fronius web interface, and is therefore not useful to us.