solarpi: Update device list a bit
Add DeviceType detection for SolarAPI for all inverters in VRM
where the device type was not yet in the list.

There are still gaps in this list. For inverters not in the list,
we will detect it as an unknown inverter, and the phase config
will be detected by fetching the 3PInverter collection.

https://github.com/victronenergy/venus/issues/1177