sunspec: continue on error if enough was detected
If there is a modbus error during the detection stage, but enough
of the models are present to continue, then don't fail.

Some SMA inverters behave differently to others that allow you to
read two registers past the end, and returns 0xFFFF in both of them.
This makes the scanning algorithm really efficient, but causes a
modbus error on these inverters because of reading one register
past the end.

Instead of failing completely, now we check if we've seen enough
models to start the service, even if it did error somewhere during
the scan.

https://github.com/victronenergy/venus-private/issues/308