enable limiting by default
The setting that allows enabling the limiter (for inverters other
than ABB and Fronius) is now enabled by default. Any inverter that
publishes model 123 or 704 should at least advertise the capability.

The setting is retained (for now) in case there are inverters out
there that needs it switched off. I am worried that some obscure
old SMA inverter is waiting in the wings.