Fix problem with duplicate UniqueId 16777215
When the UniqueId is not programmed on a PV inverter, it has the ID
16777215 (0xFFFFFF). If there is more than one of these at the same
site, all but one of them are always ignored.

Therefore, where there is more than one with this problem, use the
enumeration order of the devices as part of the identifier. Since
they always enumerate in the same order, this will allow them to
show up and settings for them to be stored.

https://github.com/victronenergy/venus/issues/1036