sunspec limiter: Slightly simplify logic
When a request is received to write a power limit, but we are
already in the middle of another request, then we would set
a variable to remember to write the limit later. This variable
was then only reset once the write completed, which is unnecessary.
We can reset it the moment the write is done. This avoids possible
bugs later on.