solaredge: Avoid dbus jitter reading maxpower
After getting the maxpower from the inverter, it was
immediately sent to dbus as the powerlimit, instead of leaving
it alone and letting sunspec_updater set these values based on
settings.

The result was that after the inverter was detected,
limiting would briefly be enabled before settings took
effect, which might disable it again.